started implemented symptoms screen
