+++ CityAtoms/CityAtoms/Assets.xcassets/iconSymptomFeverBlue.imageset/Contents.json
@@ -1,0 +1,23 @@
+{
+  "images" : [
+    {
+      "idiom" : "universal",
+      "filename" : "iconSymptomFeverBlue.png",
+      "scale" : "1x"
+    },
+    {
+      "scale" : "2x",
+      "idiom" : "universal",
+      "filename" : "iconSymptomFeverBlue@2x.png"
+    },
+    {
+      "filename" : "iconSymptomFeverBlue@3x.png",
+      "scale" : "3x",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  }
+}
