+++ CityAtoms/CityAtoms/Assets.xcassets/iconSymptomFeverGrey.imageset/Contents.json
@@ -1,0 +1,23 @@
+{
+  "images" : [
+    {
+      "scale" : "1x",
+      "idiom" : "universal",
+      "filename" : "iconSymptomFeverGrey.png"
+    },
+    {
+      "scale" : "2x",
+      "idiom" : "universal",
+      "filename" : "iconSymptomFeverGrey@2x.png"
+    },
+    {
+      "filename" : "iconSymptomFeverGrey@3x.png",
+      "scale" : "3x",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  }
+}
